Add test for attention with cxn
